Update the scoring model
This currently takes all of the criteria, applies a weight,
and uses the WebOfTrust reputation as a scalar.

This addresses the issue of the site not being flagged in
the opensources data, since their fact checking team has
limited resources and can't evaluate every site on the internet.

The downside is that using the WebOfTrust data as a multiplicative
scalar means that social/reputation data weighs more heavily than
the purely objective factors, like whether the site uses HTTPS
and practices decent editorial skills.  If we can add enough
criteria to flag content, then we can factor out the WebOfTrust
reputation and make it a smaller percentage of the total score.